Fix use crate statements after move.
