+++ src/kani-compiler/src/codegen_cprover_gotoc/codegen/place.rs
@@ -6,8 +6,8 @@
 //! in [codegen_place] below.
 
 use super::typ::TypeExt;
-use crate::utils::slice_fat_ptr;
-use crate::GotocCtx;
+use crate::codegen_cprover_gotoc::utils::slice_fat_ptr;
+use crate::codegen_cprover_gotoc::GotocCtx;
 use cbmc::goto_program::{Expr, Type};
 use rustc_hir::Mutability;
 use rustc_middle::ty::layout::LayoutOf;
