+++ src/kani-compiler/src/codegen_cprover_gotoc/context/current_fn.rs
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-use crate::GotocCtx;
+use crate::codegen_cprover_gotoc::GotocCtx;
 use cbmc::goto_program::Stmt;
 use rustc_middle::mir::BasicBlock;
 use rustc_middle::mir::Body;
